+++ NEWS.md
@@ -1,3 +1,8 @@
+# Version (development version)
+
+ * ...
+
+
 # Version 0.17.0 [2025-04-15]
 
 ## New Features