+++ NEWS.md
@@ -1,3 +1,8 @@
+# Version (development version)
+
+ * ...
+
+
 # Version 0.18.0 [2025-05-06]
 
 ## New Features